requirements.txt
aiogram==3.20.0
aiohttp
python-dotenv==1.1.0